requirements.txt
numpy
scipy
mne
torch
scikit-learn
matplotlib
PyYAML
